Revert to using local imports instead of getting them from github. #33
It turned out there was no merit in doing this.

One thing that might be confusing is that we no longer import fox as
"fox", but as "github.com/e-gov/fox". This, while it might seem so,
is not an external import (the missing "src/" directory being the key
indicator of that -- if we delete the folder in github.com, "go get" will
not recover the deleted files).

This confusion came about because we moved local packages into
canonical github.com/e-gov, which meant changes in config files,
scripts and src code.