Use the path of the url request in haproxy ACL, because port changes
The request from inside static_web (see docker-compose.yml)
to fox_LB (LB stands for load balancer, see fox_LB/haproxy.cfg) goes
through a nodejs proxy that uses CORS (changeOrigin is set to true
in static/properties.sample.json). This means that the request
port is changed and we have to rely on the request's URL path to
decide which backend to send the request.

In the case of fox service, it seems that frontend balancer can
listen on 8090, but the ACL, defined inside the balancer (see the
fox_LB/haproxy.cfg), uses the new port assigned due to CORS. #33